# packages/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LinguaQuiz API",
    "description": "Language learning quiz backend with automated spaced repetition",
    "version": "4.0.0"
  },
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Register User",
        "operationId": "register_user_api_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegistration"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Login User",
        "operationId": "login_user_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/delete-account": {
      "delete": {
        "tags": ["Authentication"],
        "summary": "Delete Account",
        "operationId": "delete_account_api_auth_delete_account_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/auth/profile": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Get User Profile",
        "operationId": "get_user_profile_api_auth_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/user/progress": {
      "get": {
        "tags": ["Progress"],
        "summary": "Get User Progress",
        "operationId": "get_user_progress_api_user_progress_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "list_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "List Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserProgressResponse"
                  },
                  "title": "Response Get User Progress Api User Progress Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Progress"],
        "summary": "Save User Progress",
        "operationId": "save_user_progress_api_user_progress_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgressUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tts/synthesize": {
      "post": {
        "tags": ["Text-to-Speech"],
        "summary": "Synthesize Speech",
        "operationId": "synthesize_speech_api_tts_synthesize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TTSRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TTSResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/tts/languages": {
      "get": {
        "tags": ["Text-to-Speech"],
        "summary": "Get Tts Languages",
        "operationId": "get_tts_languages_api_tts_languages_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TTSLanguagesResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProgressUpdateRequest": {
        "properties": {
          "vocabularyItemId": {
            "type": "string",
            "title": "Vocabularyitemid"
          },
          "level": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 0.0,
            "title": "Level"
          },
          "queuePosition": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Queueposition"
          },
          "correctCount": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Correctcount"
          },
          "incorrectCount": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Incorrectcount"
          },
          "consecutiveCorrect": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Consecutivecorrect"
          }
        },
        "type": "object",
        "required": [
          "vocabularyItemId",
          "level",
          "queuePosition",
          "correctCount",
          "incorrectCount",
          "consecutiveCorrect"
        ],
        "title": "ProgressUpdateRequest"
      },
      "TTSLanguagesResponse": {
        "properties": {
          "available": {
            "type": "boolean",
            "title": "Available"
          },
          "supportedLanguages": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Supportedlanguages"
          }
        },
        "type": "object",
        "required": ["available", "supportedLanguages"],
        "title": "TTSLanguagesResponse"
      },
      "TTSRequest": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Text"
          },
          "language": {
            "type": "string",
            "pattern": "^(en|de|ru|es)$",
            "title": "Language"
          }
        },
        "type": "object",
        "required": ["text", "language"],
        "title": "TTSRequest"
      },
      "TTSResponse": {
        "properties": {
          "audioData": {
            "type": "string",
            "title": "Audiodata"
          },
          "contentType": {
            "type": "string",
            "title": "Contenttype",
            "default": "audio/mpeg"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "language": {
            "type": "string",
            "title": "Language"
          }
        },
        "type": "object",
        "required": ["audioData", "text", "language"],
        "title": "TTSResponse"
      },
      "TokenResponse": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          },
          "expires_in": {
            "type": "string",
            "title": "Expires In",
            "default": "24h"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          }
        },
        "type": "object",
        "required": ["token", "user"],
        "title": "TokenResponse"
      },
      "UserLogin": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 50,
            "minLength": 3,
            "title": "Username"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": ["username", "password"],
        "title": "UserLogin"
      },
      "UserProgressResponse": {
        "properties": {
          "vocabularyItemId": {
            "type": "string",
            "title": "Vocabularyitemid"
          },
          "level": {
            "type": "integer",
            "title": "Level"
          },
          "queuePosition": {
            "type": "integer",
            "title": "Queueposition"
          },
          "correctCount": {
            "type": "integer",
            "title": "Correctcount"
          },
          "incorrectCount": {
            "type": "integer",
            "title": "Incorrectcount"
          },
          "consecutiveCorrect": {
            "type": "integer",
            "title": "Consecutivecorrect"
          },
          "lastPracticed": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastpracticed"
          }
        },
        "type": "object",
        "required": [
          "vocabularyItemId",
          "level",
          "queuePosition",
          "correctCount",
          "incorrectCount",
          "consecutiveCorrect",
          "lastPracticed"
        ],
        "title": "UserProgressResponse"
      },
      "UserRegistration": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 50,
            "minLength": 3,
            "title": "Username"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Password"
          }
        },
        "type": "object",
        "required": ["username", "password"],
        "title": "UserRegistration"
      },
      "UserResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "is_admin": {
            "type": "boolean",
            "title": "Is Admin",
            "default": false
          }
        },
        "type": "object",
        "required": ["id", "username"],
        "title": "UserResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}